feat(api): add delete operation

+++ src/modules/writeDataToDB.ts
@@ -1,3 +1,4 @@
+import { existsSync } from 'node:fs'
 import path from 'node:path'
 
 import fs from 'fs/promises'
@@ -5,7 +6,7 @@ import fs from 'fs/promises'
 import { generateFakeData } from './generateFakeData'
 
 // Gera dados pelo faker.js e os armazena em 'db.json', presente na pasta 'data'.
-export async function writeDataToDatabase() {
+export async function writeDataToDB() {
   const fakeData = await generateFakeData()
 
   if (fakeData) {
@@ -15,8 +16,11 @@ export async function writeDataToDatabase() {
 
     try {
       await fs.mkdir(dataDirectory, { recursive: true })
-      await fs.writeFile(filePath, readableFakeData)
-      console.log('Data written to db.json successfully')
+
+      if (existsSync('db.json')) {
+        await fs.writeFile(filePath, readableFakeData)
+        console.log('Data written to db.json successfully')
+      }
     } catch (error) {
       console.error('Failed to write data to db.json:', error)
     }
